recom/requirements.txt
# streamlit
pandas 
numpy
requests
Pillow
tensorflow
scikit-learn
